backend/requirements.txt
aiohttp==3.8.4
aioredis==2.0.1
aiosignal==1.3.1
aiosmtplib==2.0.2
alembic==1.11.1
anyio==3.7.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.2
attrs==23.1.0
audioread==3.0.1
autopep8==2.0.2
Babel==2.12.1
backcall==0.2.0
bcrypt==4.0.1
bleach==6.0.0
blinker==1.6.2
build==0.10.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.4
cloudinary==1.33.0
colorama==0.4.6
comm==0.1.4
compose==1.4.8
contourpy==1.1.0
cryptography==41.0.1
cycler==0.11.0
dataclasses-json==0.6.2
datasets==2.14.7
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
distlib==0.3.6
distro==1.8.0
dnspython==2.3.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.20.1
ecdsa==0.18.0
email-validator==1.3.0
executing==1.2.0
Faker==18.13.0
fastapi==0.100.0
fastapi-limiter==0.1.5
fastapi-mail==1.3.1
fastjsonschema==2.18.0
filelock==3.12.2
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
flatbuffers==23.5.26
fonttools==4.42.1
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.10.0
gast==0.5.4
google-auth==2.23.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.59.0
h11==0.14.0
h5py==3.9.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.19.3
idna==3.4
imagesize==1.4.1
importlib-metadata==6.8.0
iniconfig==2.0.0
ipykernel==6.25.0
ipython
isoduration==20.11.0
itsdangerous==2.1.2
jaraco.classes==3.2.3
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.4
jupyterlab-pygments==0.2.2
jupyterlab_server==2.24.0
keras
keyring==24.2.0
kiwisolver==1.4.4
langchain==0.0.336
langsmith==0.0.64
lazy_loader==0.3
libclang==16.0.6
libgravatar==1.0.4
librosa==0.10.1
llvmlite==0.41.1
lxml==4.9.3
Mako==1.2.4
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
mistune==3.0.1
ml-dtypes==0.2.0
mongoengine==0.27.0
more-itertools==9.1.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.7.3
nbformat==5.9.2
nbimporter==0.3.4
nest-asyncio==1.5.7
networkx==3.0
notebook==7.0.2
notebook_shim==0.2.3
numba==0.58.1
numpy
oauthlib==3.2.2
openai==0.27.0
opt-einsum==3.3.0
overrides==7.4.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
paramiko==3.2.0
parso==0.8.3
passlib==1.7.4
pdfminer.six==20200517
pdfplumber==0.5.28
pickleshare==0.7.5
pika==1.3.2
Pillow==10.0.0
pkginfo==1.9.6
platformdirs==3.8.1
pluggy==1.2.0
pooch==1.8.0
postgres==4.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.24.3
psutil==5.9.5
#psycopg2==2.9.6
#psycopg2-binary==2.9.6
#psycopg2-pool==1.1
pure-eval==0.2.2
pyarrow==14.0.1
pyarrow-hotfix==0.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.19.0
pydantic==1.10.11
pydantic_core==2.1.2
Pygments==2.15.1
pymongo==4.4.0
PyMySQL==1.1.0
PyNaCl==1.5.0
pyparsing==3.0.9
PyPDF2==3.0.1
pypng==0.20220715.0
pyproject_hooks==1.0.0
pyrsistent==0.19.3
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.6
pytz==2023.3
pywin32-ctypes==0.2.2
PyYAML==6.0
pyzmq==25.1.0
readme-renderer==40.0
referencing==0.29.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
resampy==0.4.2
rfc3339-validator==0.1.4
rfc3986==2.0.0
rfc3986-validator==0.1.1
rich==13.4.2
rpds-py==0.8.8
rsa==4.9
safetensors==0.4.0
scikeras==0.11.0
scikit-learn==1.3.0
scipy==1.11.2
seaborn==0.12.2
Send2Trash==1.8.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.4.1
soxr==0.3.7
SQLAlchemy==2.0.18
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
tenacity==8.2.3
tensorboard==2.14.1
tensorboard-data-server==0.7.1
#tensorflow
#tensorflow-estimator
#tensorflow-intel
#tensorflow-io-gcs-filesystem
termcolor==2.3.0
terminado==0.17.1
texttable==1.6.7
threadpoolctl==3.2.0
tinycss2==1.2.1
tokenizers==0.15.0
tomli==2.0.1
torch
torchaudio
torchvision
tornado==6.3.2
tqdm==4.66.1
traitlets==5.9.0
twine==4.0.2
typing-inspect==0.9.0
typing_extensions==4.7.1
uri-template==1.3.0
urllib3==1.26.16
utils==1.0.1
uvicorn==0.22.0
virtualenv==20.23.1
Wand==0.6.13
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.3.6
wrapt==1.14.1
xxhash==3.4.1
yarl==1.9.2
zipp==3.15.0